Added all local changes to usbtiny 1.5. This includes tinytest, tinytest-861 and keyboard-spoofer. Also included are changes to the core usbtiny library.
TinyTest is a simple Attiny2313 proof of concept.

TinyTest-861 is a demonstration of running usbtiny on an ATtiny861.

Keyboard-Spoofer is a full on USB HID Keyboard implementation that types "hello world" every few seconds.

The usbtiny library has been modified to add support for one interrupt IN endpoint. This required changes to the main interrupt loop, resulting in a slightly longer version, though critical sections remain the same. Those changes have been documented in comments.